WIP refactoring the CLI structure to make to avoid dedicated batch commands and allow future feature additions

+++ processing.py
@@ -6,7 +6,7 @@ from typing import Optional
 from helpers import determine_data_dir, count_and_log
 
 
-def split_extracted(year:int, month:int, subreddit:str, delete_source:bool=False, folder:Optional[str]=None) -> None:
+def split_extracted(year:int, month:int, subreddit:str, force:bool=False, delete_source:bool=False, folder:Optional[str]=None) -> None:
     """Split extracted subreddit/year/month files further by day"""
     subreddit = subreddit.lower()
     in_sub_dn = determine_data_dir(folder, f"extracted/monthly/{subreddit}")
@@ -26,10 +26,12 @@ def split_extracted(year:int, month:int, subreddit:str, delete_source:bool=False
                     if h_out is not None:
                         h_out.close()
                     out_fp = out_sub_dn / f"{subreddit}_{day.isoformat()}"
-                    h_out = open(out_fp, mode="w", encoding="utf-8")
-                    cur_day = day
-                    logging.info(f"Splitting comments to {out_fp}")
-                h_out.write(ln)
+                    if force is True or not out_fp.is_file():
+                        h_out = open(out_fp, mode="w", encoding="utf-8")
+                        cur_day = day
+                        logging.info(f"Splitting comments to {out_fp}")
+                if force is True or not out_fp.is_file():
+                    h_out.write(ln)
                 n = count_and_log(n)
             if h_out is not None:
                 h_out.close()
